Add files via upload
This page's purpose is to introduce what variables are with coins in a Treasure box! Users can learn what variables are by setting the amount of coins in the box. They can set it via input and then clicking the "Set Coins" button. There is also a log to show the user what coin is set to currently!